Moves config handling stuff from file manager to a new dedicated config manager.
